% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix-package.R
\docType{package}
\name{confusionMatrix}
\alias{confusionMatrix}
\alias{confusionMatrix-package}
\title{confusionMatrix: Confusion Matrix}
\description{
The goal of this package is primarily to provide an easy way to
  obtain common confusion table metrics in a tidy fashion.  The inspiration
  comes from Max Kuhn's \code{caret} package and associated function
  \code{confusionMatrix}, and the continuation of those efforts in the
  \code{yardstick} package.  Here, practically all dependencies have been
  removed except for dplyr, and results are tibbles making for easier
  document presentation, as well as the ability to peel off the statistics
  desired.

  All that is required is a vector of predicted classes and a vector of
  target classes, as that is typically what we're dealing with in such
  scenarios, i.e. predictions vs. a target variable.  These can be logical,
  integer/numeric, character, or factor, but the predictions should match the
  target in an obvious way.

  Statistics provided include:

  Accuracy and Agreement

    - Accuracy, bounds, and related

    - Cohen's Kappa

    - Corrected Rand

  Other Statistics:

    - Sensitivity

    - Specificity

    - Prevalence

    - Positive Predictive Value

    - Negative Predictive Value

    - Detection prevalence

    - Balanced Accuracy

    - F1

  Measures of Agreement/Association:

    - Phi

    - Yule's

    - Peirce's science of the method (Youden's J)

    - Jaccard
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-clark/confusionMatrix}
  \item Report bugs at \url{https://github.com/m-clark/confusionMatrix/issues}
}

}
\author{
\strong{Maintainer}: Michael Clark \email{micl@umich.edu}

}
\keyword{internal}
