% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_accuracy.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Calculate Accuracy}
\usage{
calc_accuracy(tabble)
}
\arguments{
\item{tabble}{A frequency table created with \code{\link{table}}}
}
\value{
A tibble with the corresponding statistics
}
\description{
Calculates accuracy and related metrics.
}
\details{
Used within confusion_matrix to calculate accuaracy, lower and upper
  bounds, the guessing rate and p-value of the accuracy vs. the guessing
  rate. Not really meant to be called by the user.
}
\examples{
p = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)
o = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)
calc_accuracy(table(p, o))

}
