% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Calculate various statistics from a confusion matrix}
\usage{
confusion_matrix(prediction, target, return_table = FALSE,
  dnn = c("Predicted", "target"), positive = NULL, prevalence = NULL,
  ...)
}
\arguments{
\item{prediction}{A vector of predictions}

\item{target}{A vector of target values}

\item{return_table}{Logical. Whether to return the table of \code{prediction}
vs. \code{target.} Default is \code{FALSE}.}

\item{dnn}{The row and column headers for the table returned by
\code{return_table}.}

\item{positive}{The positive class for a 2-class setting. Default is
\code{NULL}, which will result in using the first level of \code{target}.}

\item{prevalence}{Prevalence rate.  Default is \code{NULL}.}

\item{...}{Other parameters, not currently used.}
}
\value{
A list of tibble(s) with the associated statistics and possibly the
  frequency table as list column of the first element.
}
\description{
Given a vector of predictions and target values, calculate
  numerous statistics of interest.
}
\details{
This returns accuracy, agreement, and other statistics. See the
  functions below to find out more. Originally inspired by the
  \code{confusionMatrix} function from the \code{caret} package.
}
\examples{
library(confusionMatrix)

p = c(0,1,1,0)
o = c(0,1,1,1)

confusion_matrix(p, o, return_table = TRUE)

p = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)
o = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)

confusion_matrix(p, o, return_table = TRUE)

}
\references{
Kuhn, M., & Johnson, K. (2013). Applied predictive modeling.
}
\seealso{
\code{\link[caret]{confusionMatrix}}  \code{\link{calc_accuracy}}
  \code{\link{calc_agreement}} \code{\link{calc_stats}}
}
