% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Title}
\usage{
confusion_matrix(prediction, observed, return_table = FALSE,
  dnn = c("Predicted", "Observed"), positive = NULL,
  prevalence = NULL, verbose = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
