% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix-package.R
\docType{package}
\name{confusionMatrix}
\alias{confusionMatrix}
\alias{confusionMatrix-package}
\title{confusionMatrix: Confusion Matrix}
\description{
The goal of this package is primarily to provide an easy
  way to provide confusion table metrics in a tidy fashion.  The inspiration
  and a good chunk of the code borrows heavily from Max Kuhn's \code{caret}
  package and associated function \code{confusionMatrix}.  Here, practically
  all dependencies have been removed outside of dplyr, and results return
  tibbles making for easier document presentation, as well as the ability to
  peel off the statistics desired.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-clark/confusionMatrix}
  \item Report bugs at \url{https://github.com/m-clark/confusionMatrix/issues}
}

}
\author{
\strong{Maintainer}: Michael Clark \email{micl@umich.edu}

}
\keyword{internal}
